Fixes token bug in cloudsyslog_source resource
For terraform 12 go.mod was introduced, strangely requiring this module
at an older version and a different path. I think this caused older code to build,
which caused the `token` output to be blank on a sumologic_cloudsyslog_source
resource.

By removing the odd go.mod require line, and making the import in main match the
module name in go.mod (case sensitive), everything seems to build and work as expected.

Also commits go.sum as good practice.